% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{BatchExperiments}
\alias{BatchExperiments-package}
\alias{BatchExperiments}
\title{The BatchExperiments package}
\description{
Extends the BatchJobs package to run statistical experiments on
batch computing clusters.
}
\section{Additional information}{


\describe{
  \item{Homepage:}{\url{https://github.com/tudo-r/BatchExperiments}}
  \item{Wiki:}{\url{https://github.com/tudo-r/BatchExperiments/wiki}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tudo-r/BatchExperiments}
  \item Report bugs at \url{https://github.com/tudo-r/BatchExperiments/issues}
}

}
\author{
\strong{Maintainer}: Michel Lang \email{michellang@gmail.com}

Authors:
\itemize{
  \item Bernd Bischl \email{bernd_bischl@gmx.net}
  \item Olaf Mersmann \email{olafm@p-value.net}
}

}
