% Generated by roxygen2 (4.1.0.9001): do not edit by hand
\name{expandDesign}
\alias{expandDesign}
\title{Expands an Algorithm/Problem Design}
\usage{
expandDesign(x, ...)
}
\arguments{
  \item{design}{a -- list of -- \code{Design} objects.}
}
\value{
  a -- list of -- \code{dara.frame}.
}
\description{
  Expands a design parameter grid into an exhaustive \code{data.frame}.
}

