% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment.R
\name{ExperimentJob}
\alias{ExperimentJob}
\title{ExperimentJob}
\description{
You can access job properties using the \code{job} object which is optionally passed
to dynamic problem functions and algorithms. The object is a named list with the following
elements:
\describe{
  \item{\code{id} [\code{integer(1)}]:}{Job ID.}
  \item{\code{prob.id} [\code{character(1)}]:}{Problem ID.}
  \item{\code{prob.pars} [\code{list}]:}{Problem parameters as named list.}
  \item{\code{algo.id} [\code{character(1)}]:}{algo.id}{Algorithm ID.}
  \item{\code{algo.pars} [\code{list}]:}{Algorithm parameters as named list.}
  \item{\code{repl} [\code{integer(1)}]:}{Replication number of this experiment.}
  \item{\code{seed} [\code{integer(1)}]:}{Seed set right before algorithm execution.}
  \item{\code{prob.seed} [\code{integer(1)}]:}{Seed set right before generation of problem instance.}
}
}

