% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getAlgorithm}
\alias{getAlgorithm}
\title{Get algorithm from registry by id.
The object is loaded from disk.}
\usage{
getAlgorithm(reg, id)
}
\arguments{
\item{reg}{[\code{\link{ExperimentRegistry}}]\cr
Registry.}

\item{id}{[\code{character(1)}]\cr
Id of algorithm.}
}
\value{
[\code{\link{Algorithm}}].
}
\description{
Get algorithm from registry by id.
The object is loaded from disk.
}

