% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlgorithmIds.R
\name{getAlgorithmIds}
\alias{getAlgorithmIds}
\title{Get ids of algorithms in registry.}
\usage{
getAlgorithmIds(reg, ids)
}
\arguments{
\item{reg}{[\code{\link{ExperimentRegistry}}]\cr
Registry.}

\item{ids}{[code{integer}]\cr
Job ids to restrict returned algorithm ids to.}
}
\value{
[\code{character}].
}
\description{
Get algorithm ids for jobs.
}
\seealso{
Other get: 
\code{\link{getAlgorithm}()},
\code{\link{getExperimentParts}()},
\code{\link{getJobs.ExperimentRegistry}()},
\code{\link{getProblemIds}()},
\code{\link{getProblem}()}
}
\concept{get}
