% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getExperimentParts}
\alias{getExperimentParts}
\title{Get all parts required to run a single job.}
\usage{
getExperimentParts(reg, id)
}
\arguments{
\item{reg}{[\code{\link{ExperimentRegistry}}]\cr
Registry.}

\item{ids}{[\code{integer(1)}]\cr
Id of a job.}
}
\value{
[named list]. Returns the \link{Job}, \link{Problem} and \link{Instance}.
}
\description{
Get all parts required to run a single job.
}
\seealso{
Other get: \code{\link{getAlgorithmIds}};
  \code{\link{getAlgorithm}};
  \code{\link{getJobs.ExperimentRegistry}};
  \code{\link{getProblemIds}}; \code{\link{getProblem}}
}

