% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getProblem}
\alias{getProblem}
\title{Get problem from registry by id.}
\usage{
getProblem(reg, id)
}
\arguments{
\item{reg}{[\code{\link{ExperimentRegistry}}]\cr
Registry.}

\item{id}{[\code{character(1)}]\cr
Id of problem.}
}
\value{
[\code{\link{Problem}}].
}
\description{
The object is loaded from disk.
}
\seealso{
Other get: \code{\link{getAlgorithmIds}};
  \code{\link{getAlgorithm}};
  \code{\link{getJobs.ExperimentRegistry}};
  \code{\link{getProblemIds}}
}

