\name{getProblemIds}
\alias{getProblemIds}
\title{Get ids of problems in registry.}
\usage{
getProblemIds(reg, ids)
}
\arguments{
  \item{reg}{[\code{\link{ExperimentRegistry}}]\cr
  Registry.}

  \item{ids}{[code{integer}]\cr Job ids to restrict
  returned problem ids to.}
}
\value{
[\code{character}].
}
\description{
Get ids of problems in registry.
}

