% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{BatchExperiments}
\alias{BatchExperiments}
\alias{BatchExperiments-package}
\title{The BatchExperiments package}
\description{
The BatchExperiments package
}
\section{Additional information}{


\describe{
  \item{Homepage:}{\url{https://github.com/tudo-r/BatchExperiments}}
  \item{Wiki:}{\url{https://github.com/tudo-r/BatchExperiments/wiki}}
}
}

