\docType{package}
\name{BatchExperiments}
\alias{BatchExperiments}
\alias{BatchExperiments-package}
\title{The BatchExperiments package}
\description{
The BatchExperiments package
}
\section{Additional information}{
  Wiki, FAQ, examples and more on
  \url{https://github.com/tudo-r/BatchExperiments}.
}

