% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{BatchExperiments}
\alias{BatchExperiments}
\alias{BatchExperiments-package}
\title{The BatchExperiments package}
\description{
The BatchExperiments package
}
\section{Additional information}{


\describe{
  \item{Homepage:}{\url{https://github.com/tudo-r/BatchExperiments}}
  \item{Wiki:}{\url{https://github.com/tudo-r/BatchExperiments/wiki}}
}
}

