% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateProblemInstance.R
\name{generateProblemInstance}
\alias{Instance}
\alias{generateProblemInstance}
\title{Generate dynamic part of problem.}
\usage{
generateProblemInstance(reg, id)
}
\arguments{
\item{reg}{[\code{\link{ExperimentRegistry}}]\cr
Registry.}

\item{id}{[\code{character(1)}]\cr
Id of job.}
}
\value{
Dynamic part of problem.
}
\description{
Calls the dynamic problem function on the static problem part and
thereby creates the problem instance.
The seeding mechanism is identical to execution on the slave.
}

