\name{generateProblemInstance}
\alias{generateProblemInstance}
\title{Generate dynamic part of problem.}
\usage{
generateProblemInstance(reg, id)
}
\arguments{
  \item{reg}{[\code{\link{ExperimentRegistry}}]\cr
  Registry.}

  \item{id}{[\code{character(1)}]\cr Id of job.}
}
\value{
Dynamic part of problem.
}
\description{
Generate dynamic part of problem.
}

