\name{getAlgorithmIds}
\alias{getAlgorithmIds}
\title{Get ids of algorithms in registry.}
\usage{
getAlgorithmIds(reg, ids)
}
\arguments{
  \item{reg}{[\code{\link{ExperimentRegistry}}]\cr
  Registry.}

  \item{ids}{[code{integer}]\cr Job ids to restrict
  returned algorithm ids to.}
}
\value{
[\code{character}].
}
\description{
Get ids of algorithms in registry.
}

