% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperimentParts.R
\name{getExperimentParts}
\alias{getExperimentParts}
\title{Get all parts required to run a single job.}
\usage{
getExperimentParts(reg, id)
}
\arguments{
\item{reg}{[\code{\link{ExperimentRegistry}}]\cr
Registry.}

\item{id}{[\code{integer(1)}]\cr
Id of a job.}
}
\value{
[named list]. Returns the \link[BatchJobs]{Job}, \link[BatchExperiments]{Problem},
\link[BatchExperiments]{Instance} and \link[BatchExperiments]{Algorithm}.
}
\description{
Get all parts which define an \code{\link{Experiment}}.
}
\seealso{
Other get: 
\code{\link{getAlgorithm}()},
\code{\link{getAlgorithmIds}()},
\code{\link{getJobs.ExperimentRegistry}()},
\code{\link{getProblem}()},
\code{\link{getProblemIds}()}
}
\concept{get}
