% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getProblem}
\alias{getProblem}
\title{Get problem from registry by id.}
\usage{
getProblem(reg, id)
}
\arguments{
\item{reg}{[\code{\link{ExperimentRegistry}}]\cr
Registry.}

\item{id}{[\code{character(1)}]\cr
Id of problem.}
}
\value{
[\code{\link{Problem}}].
}
\description{
The object is loaded from disk.
}

