% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProblem.R
\name{getProblem}
\alias{getProblem}
\title{Get problem from registry by id.}
\usage{
getProblem(reg, id)
}
\arguments{
\item{reg}{[\code{\link{ExperimentRegistry}}]\cr
Registry.}

\item{id}{[\code{character(1)}]\cr
Id of problem.}
}
\value{
[\code{\link{Problem}}].
}
\description{
The requested object is loaded from disk.
}
\seealso{
Other get: \code{\link{getAlgorithmIds}},
  \code{\link{getAlgorithm}},
  \code{\link{getExperimentParts}},
  \code{\link{getJobs.ExperimentRegistry}},
  \code{\link{getProblemIds}}
}
