% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProblemIds.R
\name{getProblemIds}
\alias{getProblemIds}
\title{Get ids of problems in registry.}
\usage{
getProblemIds(reg, ids)
}
\arguments{
\item{reg}{[\code{\link{ExperimentRegistry}}]\cr
Registry.}

\item{ids}{[code{integer}]\cr
Job ids to restrict returned problem ids to.}
}
\value{
[\code{character}].
}
\description{
Get problem ids for jobs.
}
\seealso{
Other get: \code{\link{getAlgorithmIds}},
  \code{\link{getAlgorithm}},
  \code{\link{getExperimentParts}},
  \code{\link{getJobs.ExperimentRegistry}},
  \code{\link{getProblem}}
}
